+++ code/nel/src/net/module_l5_transport.cpp
@@ -177,7 +177,7 @@ namespace NLNET
 
 		virtual uint32 getRouteCount() const
 		{
-			return _Routes.size();
+			return (uint32)_Routes.size();
 		}
 
 		void dump(NLMISC::CLog &log) const
